requirements.txt
pyjwt==2.3.0
sanic==21.9.1
requests==2.26.0